start creating unit testing

+++ resources/views/product/show.blade.php
@@ -42,7 +42,7 @@
         <div class="col-sm-6 col-sm-offset-3 text-center">
             <a href="{{url('/')}}" class="btn btn-primary">Show List</a>
             <a href="{{url('product/'. $product->id . '/edit')}}" class="btn btn-primary">Edit</a>
-            {{ Form::open(array('route' => array('product.destroy', $product['id']), 'method' => 'delete', 'class'=>'inline-block')) }}
+            {{ Form::open(array('route' => array('product.destroy', $product['id']), 'method' => 'delete', 'class'=>'inline-block', 'onsubmit'=>'return confirm("Do you want to delete this product?")')) }}
             {!! Form::submit('Delete', ["class"=>"btn btn-danger", "href"=>"#"]) !!}
             {{ Form::close() }}
         </div>
